[Chore #155855873] Renames file according to functionality
