+++ src/pages/ConfirmRegistrationPage.js
