+++ src/pages/SingleEventPage.js
